backend_v2/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart
python-dotenv
chromadb
sentence_transformers
PyPDF2
psycopg2-binary
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
aiohttp==3.9.1
urllib3==2.0.7
